Clean up codebase
